requirements.txt
Flask
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
sqlalchemy
pymysql
python-dotenv
gspread
google-auth
